+++ src/components/incidence/IncidenceLabel.vue
